Covid 19 & X-Ray- Project Proposal
This project serves the health sector to save a lot of effort and money.
In this project, we will see by using Deep learning if the case is for injured or
uninjured. The idea of the project is to know if the X-ray images for the patient are
infected or not. , based on the X-ray pictures.